Correct default flavor size
In the previous commit, this was changed because the 8GB + 50GB RAM/Disk
respectively made it a bit tight for most workloads.
I was meant to go one step up (tiny @ 16/100), but instead went to small
which is 128GB RAM per machine.

This meant, to deploy a cluster, the user would need 0.5TB of
RAM...let's adjust it back to something sensible