Add ingress controller, disabled by default
Adds the config for the ingress controller, this will steer users to not
allocate a FIP per service, makes TLS trivial (using certbot or similar)
and avoids duplicating effort because we use the StackHPC charts.

This disabled by default, pending documentation and additional testing.
But I want to ensure the correct values is captured whilst we're here.